backend/mycrud/views.py
from rest_framework import generics
from mycrud.models import Note
from mycrud.serializers import NoteSerializer


class NoteView(generics.ListCreateAPIView):
    queryset = Note.objects.all()
    serializer_class = NoteSerializer


class NoteDetailView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Note.objects.all()
    serializer_class = NoteSerializer
