backend/requirements.txt
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0